feat: updated delete cart product
if you have a cart product with quantity > 1, and want to decrement
it, you can now use DeleteCartProduct with oneQuantity == true.

if you have a cart product with quantity > 1, and want to delete all
references to this product in a cart, you can now use DeleteCartProduct
with oneQuantity == false, which will delete the whole cartProduct
object.